Move the external gateways tests to a different file.
We collect all the "external gateway" kind tests in a different file for
better discovery / navigation
